move TemplatesMixin back into templates.py
