--- bocadillo/deprecated/__init__.py
@@ -1,1 +1,0 @@
-from .templates import TemplatesMixin
