+++ bocadillo/templates.py
@@ -1,12 +1,9 @@
 from contextlib import contextmanager, suppress
 from typing import Any, Optional, cast
 
-try:
-    import jinja2
-except ImportError as exc:  # pragma: no cover
-    jinja2 = None  # type: ignore
-else:
-    from jinja2 import Environment, FileSystemLoader, Template
+from jinja2 import Environment, FileSystemLoader, Template
+
+from .deprecation import ReplacedBy
 
 
 DEFAULT_TEMPLATES_DIR = "templates"
@@ -15,7 +12,7 @@ DEFAULT_TEMPLATES_DIR = "templates"
 class Templates:
     """This class provides templating capabilities.
 
-    This is a light wrapper around [jinja2](http://jinja.pocoo.org/docs) and using it requires to install Bocadillo using the `[templates]` extra.
+    This is a light wrapper around [jinja2](http://jinja.pocoo.org/docs).
 
     See also [Templates](../guides/agnostic/templates.md) for detail.
 
@@ -46,12 +43,6 @@ class Templates:
         directory: str = DEFAULT_TEMPLATES_DIR,
         context: dict = None,
     ):
-        if jinja2 is None:
-            raise ImportError(
-                "`jinja2` is not installed. "
-                "Did you install Bocadillo using `bocadillo[templates]`?"
-            ) from exc
-
         if context is None:
             context = {}
 
@@ -138,3 +129,78 @@ class Templates:
         """
         template = self._environment.from_string(source=source)
         return template.render(*args, **kwargs)
+
+
+# DEPRECATED: 0.13.0
+
+_REPLACED_BY = ReplacedBy(
+    since="0.12",
+    removal="0.13",
+    obj_root="bocadillo.templates.",
+    doc_root="./templates.md",
+)
+
+
+@_REPLACED_BY("Templates", "#templates")
+class TemplatesMixin:
+    """Provide templating capabilities to an application class."""
+
+    def __init__(self, templates_dir: str = DEFAULT_TEMPLATES_DIR, **kwargs):
+        super().__init__(**kwargs)
+        self._templates = Templates(
+            directory=templates_dir, context=self.get_template_globals()
+        )
+
+    def get_template_globals(self) -> dict:
+        return {}
+
+    @property  # type: ignore
+    @_REPLACED_BY("Templates.directory", "#templates")
+    def templates_dir(self) -> Optional[str]:
+        """The path where templates are searched for, or `None` if not set.
+
+        This is built from the `templates_dir` parameter.
+        """
+        return self._templates.directory
+
+    @templates_dir.setter  # type: ignore
+    @_REPLACED_BY("Templates.directory", "#templates")
+    def templates_dir(self, templates_dir: str):
+        self._templates.directory = templates_dir
+
+    @_REPLACED_BY("Templates.render", "#render")
+    async def template(self, name_: str, *args: dict, **kwargs: Any) -> str:
+        """Render a template asynchronously.
+
+        Can only be used within `async` functions.
+
+        # Parameters
+        name (str):
+            Name of the template, located inside `templates_dir`.
+            The trailing underscore avoids collisions with a potential
+            context variable named `name`.
+        *args (dict):
+            Context variables to inject in the template.
+        **kwargs (any):
+            Context variables to inject in the template.
+        """
+        return await self._templates.render(name_, *args, **kwargs)
+
+    @_REPLACED_BY("Templates.render_sync", "#render-sync")
+    def template_sync(self, name_: str, *args: dict, **kwargs: Any) -> str:
+        """Render a template synchronously.
+
+        For parameters, see [.template()](#template).
+        """
+        return self._templates.render_sync(name_, *args, **kwargs)
+
+    @_REPLACED_BY("Templates.render_string", "#render-string")
+    def template_string(self, source: str, *args: dict, **kwargs: Any) -> str:
+        """Render a template from a string (synchronous).
+
+        # Parameters
+        source (str): a template given as a string.
+        
+        For other parameters, see [.template()](#template).
+        """
+        return self._templates.render_string(source, *args, **kwargs)
